+++ bydbctl/internal/cmd/rest.go
@@ -18,9 +18,12 @@
 package cmd
 
 import (
+	"crypto/tls"
+	"crypto/x509"
 	"encoding/json"
 	"fmt"
 	"io"
+	"os"
 	"strings"
 	"time"
 
@@ -290,7 +293,7 @@ func yamlPrinter(index int, _ reqBody, body []byte) error {
 	return nil
 }
 
-func rest(pfn paramsFn, fn reqFn, printer printer) (err error) {
+func rest(pfn paramsFn, fn reqFn, printer printer, enableTLS bool, insecure bool, grpcCert string) (err error) {
 	var requests []reqBody
 	if pfn == nil {
 		requests = []reqBody{{}}
@@ -302,7 +305,26 @@ func rest(pfn paramsFn, fn reqFn, printer printer) (err error) {
 	}
 
 	for i, r := range requests {
-		req := resty.New().R()
+		client := resty.New()
+		if enableTLS {
+			// #nosec G402
+			config := tls.Config{
+				InsecureSkipVerify: insecure,
+			}
+			if grpcCert != "" {
+				cert, err := os.ReadFile(grpcCert)
+				if err != nil {
+					return err
+				}
+				certPool := x509.NewCertPool()
+				if !certPool.AppendCertsFromPEM(cert) {
+					return errors.New("failed to add server's certificate")
+				}
+				config.RootCAs = certPool
+			}
+			client.SetTLSClientConfig(&config)
+		}
+		req := client.R()
 		resp, err := fn(request{
 			reqBody: r,
 			req:     req,